+++ app/packages/dataset/src/Dataset.tsx
@@ -12,7 +12,6 @@ import {
   Dataset as CoreDataset,
   useDatasetLoader,
   usePreLoadedDataset,
-  useSavedViewLoader,
   ViewBar,
 } from "@fiftyone/core";
 import { useRecoilValue, useSetRecoilState } from "recoil";
